+++ src/Sitra/ApiClient/Description/Metadata.php
@@ -115,6 +115,8 @@ class Metadata
 
                 if (is_array($data)) {
                     $metadata[$name] = json_encode($data);
+                } else if (empty($data)) {
+                    throw new InvalidMetadataFormatException();
                 }
             }
         }